+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - fix: Correctly cross year boundries
 - feat: Add convenience methods `isHoliday` and `isSpecialDay`
 - feat: Returns special days sorted in correct date order
+- feat: Add `sjomanna` special day ("Sjómannadagurinn")
 - feat: Shorten name/description of `jun17` to "Þjóðhátíðardagurinn"
 - fix: Handle `0` value year and month arguments correctly
 - fix: Correctly handle rímspilliar shifting `bonda` and `konu` days